requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.3
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
marshmallow==3.17.0
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
typing-extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1
PyJWT==2.4.0
pytest==7.1.2
atomicwrites==1.4.1
colorama==0.4.5
iniconfig==1.1.1
packaging==21.3
pip==22.1.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
setuptools==63.2.0
tomli==2.0.1
wheel==0.37.1
